# hcq optimize enqueue time