`BEAM=2 NV=1 python3 examples/llama.py --temperature 0 --count 10 --prompt "Hello." --timing --shard 4`

```
enqueue in   4.48 ms
total   9.55 ms, 104.73 tok/s, 1466.57 GB/s, param 1412.44 GB/s
```